requirements.txt
altgraph==0.17.2
future==0.18.2
pefile==2021.9.3
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PySimpleGUI==4.49.0
pywin32-ctypes==0.2.0
PyYAML==5.4.1
